+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * 【core】       修改错别字（pr#568@Github）
 * 【core】       增加DateUtil.parseCST方法（issue#570@Github）
 * 【core】       增加defaultIfEmpty方法
+* 【crypto】     修改bigIntToFixexLengthBytes为bigIntToFixedLengthBytes（pr#575@Github）
 
 ### Bug修复
 * 【all】        修复阶乘计算错误bug（issue#I12XE4@Gitee）